refactor: benchmark code generic
