+++ xtask/bench/benches/analyzer.rs
@@ -1,4 +1,4 @@
-use criterion::{criterion_group, criterion_main, Criterion};
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use std::collections::HashMap;
 use xtask_bench::{Parse, TestCase};
 
@@ -9,39 +9,38 @@ fn bench_analyzer(criterion: &mut Criterion) {
     let mut libs = vec![];
     libs.extend(all_suites.values().flat_map(|suite| suite.lines()));
 
+    let mut group = criterion.benchmark_group("analyzer");
+
     for lib in libs {
         let test_case = TestCase::try_from(lib);
 
         match test_case {
             Ok(test_case) => {
-                let parse = Parse::try_from_case(&test_case).expect("Supported language");
-
                 let code = test_case.code();
-
-                let mut group = criterion.benchmark_group("analyzer");
                 group.throughput(criterion::Throughput::Bytes(code.len() as u64));
+                group.bench_with_input(
+                    BenchmarkId::from_parameter(test_case.filename()),
+                    code,
+                    |b, _| {
+                        let parse = Parse::try_from_case(&test_case).expect("Supported language");
 
-                let parsed = parse.parse();
+                        let parsed = parse.parse();
 
-                match parsed.analyze() {
-                    None => {
-                        continue;
-                    }
-                    Some(analyze) => {
-                        group.bench_function(test_case.filename(), |b| {
-                            b.iter(|| {
+                        match parsed.analyze() {
+                            None => {}
+                            Some(analyze) => b.iter(|| {
                                 analyze.analyze();
                                 criterion::black_box(());
-                            })
-                        });
-                    }
-                }
-
-                group.finish();
+                            }),
+                        }
+                    },
+                );
             }
             Err(e) => println!("{:?}", e),
         }
     }
+
+    group.finish();
 }
 
 criterion_group!(analyzer, bench_analyzer);
