+++ xtask/bench/benches/json_parser.rs
@@ -1,7 +1,6 @@
-use biome_rowan::NodeCache;
-use criterion::{criterion_group, criterion_main, BatchSize, BenchmarkId, Criterion};
+use criterion::{criterion_group, criterion_main, Criterion};
 use std::collections::HashMap;
-use xtask_bench::{Parse, TestCase};
+use xtask_bench::{bench_parser_group, TestCase};
 
 fn bench_json_parser(criterion: &mut Criterion) {
     let mut all_suites = HashMap::new();
@@ -9,42 +8,18 @@ fn bench_json_parser(criterion: &mut Criterion) {
     let mut libs = vec![];
     libs.extend(all_suites.values().flat_map(|suite| suite.lines()));
 
+    let mut group = criterion.benchmark_group("json_parser");
     for lib in libs {
         let test_case = TestCase::try_from(lib);
 
         match test_case {
             Ok(test_case) => {
-                let parse = Parse::try_from_case(&test_case).expect("Supported language");
-
-                let code = test_case.code();
-
-                let mut group = criterion.benchmark_group("json_parser");
-                group.throughput(criterion::Throughput::Bytes(code.len() as u64));
-
-                group.bench_function(BenchmarkId::new(test_case.filename(), "uncached"), |b| {
-                    b.iter(|| {
-                        criterion::black_box(parse.parse());
-                    })
-                });
-                group.bench_function(BenchmarkId::new(test_case.filename(), "cached"), |b| {
-                    b.iter_batched(
-                        || {
-                            let mut cache = NodeCache::default();
-                            parse.parse_with_cache(&mut cache);
-                            cache
-                        },
-                        |mut cache| {
-                            criterion::black_box(parse.parse_with_cache(&mut cache));
-                        },
-                        BatchSize::SmallInput,
-                    )
-                });
-
-                group.finish();
+                bench_parser_group(&mut group, test_case);
             }
             Err(e) => println!("{:?}", e),
         }
     }
+    group.finish();
 }
 
 criterion_group!(json_parser, bench_json_parser);
